--- a/.github/workflows/flakes.yml
+++ b/.github/workflows/flakes.yml
@@ -15,7 +15,7 @@ jobs:
     - uses: cachix/cachix-action@v11
       with:
         name: flora-pm
-        authToken: '${{ secrets.CACHIX_AUTH_TOKEN }}'
+        authToken: ${{ secrets.CACHIX_AUTH_TOKEN }}
         extraPullNames: nix-community
     - run: nix build
     - run: nix flake check