--- a/.github/workflows/flakes.yml
+++ b/.github/workflows/flakes.yml
@@ -1,0 +1,21 @@
+name: "Nix flakes check"
+on:
+  pull_request:
+  push:
+    branches: ["main", "development"]
+jobs:
+  tests:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v3
+    - uses: cachix/install-nix-action@v18
+      with:
+        extra_nix_config: |
+          access-tokens = github.com=${{ secrets.GITHUB_TOKEN }}
+    - uses: cachix/cachix-action@v11
+      with:
+        name: flora-pm
+        authToken: '${{ secrets.CACHIX_AUTH_TOKEN }}'
+        extraPullNames: nix-community
+    - run: nix build
+    - run: nix flake check