--- a/src/Flora/Import/Package/Bulk.hs
+++ b/src/Flora/Import/Package/Bulk.hs
@@ -2,19 +2,15 @@
 
 module Flora.Import.Package.Bulk (importAllFilesInDirectory, importAllFilesInRelativeDirectory) where
 
-import Control.Concurrent (getNumCapabilities, modifyMVar, newMVar)
-import Control.Concurrent.Async (async, wait)
-import Data.Foldable (traverse_)
-import Data.Function
+import Control.Concurrent (modifyMVar, newMVar, readMVar)
+import Control.Monad ((>=>), when)
 import Data.List (isSuffixOf)
 import Effectful
 import Effectful.Log qualified as Log
 import Effectful.PostgreSQL.Transact.Effect (DB, getPool, runDB)
 import Effectful.Time
 import Log (Logger, defaultLogLevel)
-import Streaming (chunksOf)
-import Streaming.Prelude (Of, Stream)
-import Streaming.Prelude qualified as Str
+import qualified Streamly.Prelude as S
 import System.Directory qualified as System
 import System.FilePath
 
@@ -33,49 +29,35 @@ importAllFilesInRelativeDirectory appLogger user dir = do
 importAllFilesInDirectory :: (DB :> es, IOE :> es) => Logger -> UserId -> FilePath -> Eff es ()
 importAllFilesInDirectory appLogger user dir = do
   pool <- getPool
-  parallelWorkers <- liftIO getNumCapabilities
-  let chunkSize = 400
   countMVar <- liftIO $ newMVar @Int 0
-  findAllCabalFilesInDirectory dir
-    & parMapM parallelWorkers (runEff . runDB pool . runCurrentTimeIO . Log.runLog "flora-jobs" appLogger defaultLogLevel . loadAndExtractCabalFile user)
-    & chunksOf chunkSize
-    & Str.mapped Str.toList
-    & Str.mapM_ (persistChunk countMVar)
+  liftIO $ System.createDirectoryIfMissing True dir
+  liftIO . putStrLn $ "🔎  Searching cabal files in " <> dir
+  liftIO $ S.drain $ S.fromParallel $ S.mapM (processFile pool countMVar) $ findAllCabalFilesInDirectory dir
   Update.refreshLatestVersions >> Update.refreshDependents
   where
-    persistChunk countMvar chunk = do
-      let size = length chunk
-      newCount <- liftIO $ modifyMVar countMvar (\c -> pure (c + size, c + size))
-      traverse_ persistImportOutput chunk
-      liftIO . putStrLn $ "✅ Processed " <> show newCount <> " new cabal files"
+    processFile' pool =
+        runEff
+      . runDB pool
+      . runCurrentTimeIO
+      . Log.runLog "flora-jobs" appLogger defaultLogLevel
+      . (loadAndExtractCabalFile user >=> persistImportOutput)
+    processFile pool countMVar p = do
+      processFile' pool p
+      newCount <- modifyMVar countMVar (\c -> pure (c + 1, c + 1))
+      when (newCount `mod` 400 == 0)$
+        displayStats countMVar
+    displayStats countMVar = do
+      currentCount <- readMVar countMVar
+      liftIO . putStrLn $ "✅ Processed " <> show currentCount <> " new cabal files"
 
-{-| Finds all cabal files in the provided directory recursively
- Hits are written to the output channel as they are found, so it should be possible to process
- large amounts of Cabal files efficiently
--}
-findAllCabalFilesInDirectory
-  :: IOE :> es
-  => FilePath
-  -> Stream (Of FilePath) (Eff es) ()
-findAllCabalFilesInDirectory workdir = do
-  liftIO . putStrLn $ "🔎  Searching cabal files in " <> workdir
-  liftIO $ System.createDirectoryIfMissing True workdir
-  inspectDir workdir
+findAllCabalFilesInDirectory :: FilePath -> S.ParallelT IO FilePath
+findAllCabalFilesInDirectory workdir = S.concatMapM traversePath $ S.fromList [workdir]
   where
-    inspectDir dir = liftIO (System.listDirectory dir) >>= traverse_ (inspectItem dir)
-    inspectItem dir item = do
-      let fullPath = dir </> item
-      isDir <- liftIO $ System.doesDirectoryExist fullPath
+    traversePath p = do
+      isDir <- liftIO $ System.doesDirectoryExist p
       case isDir of
-        True -> inspectDir fullPath
-        False | ".cabal" `isSuffixOf` fullPath -> Str.yield fullPath
-        _ -> pure ()
-
--- | Replaces each element of a stream with the result of an action, processing elements of the stream concurrently
-parMapM :: IOE :> es => Int -> (a -> IO b) -> Stream (Of a) (Eff es) () -> Stream (Of b) (Eff es) ()
-parMapM concurrentActions f str =
-  Str.mapM (liftIO . async . f) str
-    & chunksOf concurrentActions
-    & Str.mapped Str.toList
-    & Str.mapM (liftIO . traverse wait)
-    & Str.concat
+        True -> do
+          entries <- System.listDirectory p
+          return $ S.concatMapM (traversePath . (p </>)) $ S.fromList entries
+        False | ".cabal" `isSuffixOf` p -> return $ S.fromPure p
+        _ -> return S.nil