## Proposed changes

It's actually a draft since it has the following shortcomings:

- [ ] A memory leak (13.5 GB at the end)
- [ ] Very low parallelization (x6 on 20-core CPU)
- [ ] Outputs is interleaved

Results (on a 20-core CPU):
- `before`: 4745.19s user 916.19s system 144% cpu 57:59.78 total
- `after`: 6869.13s user 1557.62s system 619% cpu 22:39.85 total

## Contributor checklist

- [ ] My PR is related to \<insert ticket number> 
- [x] I have read and understood the [CONTRIBUTING guide](https://github.com/flora-pm/flora-server/blob/development/CONTRIBUTING.md)
- [ ] I have inserted my change and a link to this PR in the [CHANGELOG](https://github.com/flora-pm/flora-server/blob/development/CHANGELOG.md)