@tchoutri 2022-11-22
@blackheaven regarding interleaved output it's alright