@blackheaven 2022-11-22
I figured out it was already a defect.

It'll take a proper logger and an effect change to be fixed.

Regarding nix overrides/layers, there are incompatible derivations, I have to figure out how to override it locally.